Add some missing documentation

+++ GoTrue/src/commonMain/kotlin/io/github/jan/supabase/gotrue/mfa/MfaChallenge.kt
@@ -9,6 +9,6 @@ data class MfaChallenge(val id: String) {
 
     @SerialName("expires_at") private val expiresAtSeconds: Long = 0
     val expiresAt: Instant
-        get() = Instant.fromEpochSeconds(expiresAtSeconds.toLong())
+        get() = Instant.fromEpochSeconds(expiresAtSeconds)
 
 }
